+++ mobile/app/course/[slug]/revision.tsx
@@ -1,0 +1,3 @@
+import { CourseRevisionScreen } from '../../../src/screens/courses/CourseRevisionScreen';
+
+export default CourseRevisionScreen;